requirements.txt
# General Utilities
tqdm
requests
urllib3
PyYAML
python-dotenv

# Data Science & ML
numpy
pandas
matplotlib
scikit-learn
h5py
scipy

# Image Processing
opencv-python
Pillow

# Deep Learning
torch
torchvision
datasets
kagglehub
paddlepaddle
paddleocr
realesrgan

# Other
reportlab
Brotli
Unidecode
ipykernel
kagglesdk
pycryptodomex
mysql-connector-python
ultralytics
dotenv
